backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
PyJWT==2.1.0
pytest==7.3.1
pytest-django==4.5.2
pytest-pythonpath==0.7.3
python-dotenv==1.0.0 